+++ Changelog.rst
@@ -16,7 +16,7 @@ New features
   Modularized and simplified all code (C and Python) related to block ciphers.
   Pycryptodome is now free of CPython extensions.
 * Add support for CI in Windows via Appveyor.
-* RSA key generation more closely follows FIPS 186-4 (though it is
+* RSA and DSA key generation more closely follows FIPS 186-4 (though it is
   not 100% compliant).
 
 Resolved issues